@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import stark.dataworks.boot.ExceptionLogger;
 import stark.dataworks.boot.web.ServiceResponse;
 
-import javax.validation.ConstraintViolationException;
+import jakarta.validation.ConstraintViolationException;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler
@@ -16,5 +16,12 @@ public class GlobalExceptionHandler
     {
         ExceptionLogger.logExceptionInfo(e);
         return ServiceResponse.buildErrorResponse(-100, e.getMessage());
+    }
+
+    @ExceptionHandler({Exception.class})
+    public ServiceResponse<?> handleException(Exception e)
+    {
+        ExceptionLogger.logExceptionInfo(e);
+        return ServiceResponse.buildErrorResponse(-200, e.getMessage());
     }
 }
